# src/Api.Core.SolarMeasurementsProxy/AutoGeneratedProxyClasses/Models/Measurement.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Api.Core.SolarMeasurementsProxy.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class Measurement
    {
        /// <summary>
        /// Initializes a new instance of the Measurement class.
        /// </summary>
        public Measurement()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the Measurement class.
        /// </summary>
        public Measurement(System.DateTime? timestamp = default(System.DateTime?), int? sensorId = default(int?), double? value = default(double?))
        {
            Timestamp = timestamp;
            SensorId = sensorId;
            Value = value;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "timestamp")]
        public System.DateTime? Timestamp { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "sensorId")]
        public int? SensorId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "value")]
        public double? Value { get; set; }

    }
}