# src/Api.Core.SolarMeasurementsProxy/AutoGeneratedSwaggerDefinition/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SolarMeasurements Core API",
    "description": "A simple example API for retrieving measurements from Solar sites",
    "license": {
      "name": "Licensed under MIT Licence",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "paths": {
    "/api/v1.0/SolarMeasurements": {
      "get": {
        "tags": [
          "SolarMeasurements"
        ],
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": ""
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": ""
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "hourly",
                "daily",
                "weekly",
                "monthly"
              ],
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Measurement"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1.0/SolarMeasurements/{timestamp}": {
      "get": {
        "tags": [
          "SolarMeasurements"
        ],
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Measurement"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Measurement": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "sensorId": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}